Move trigram generation code from util into the package
The code to generate normalized trigrams from text may be useful to
users but was not included in the package distribution. It is now
available in the icegrams package alongside the Ngrams class.
